% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_performance.R
\name{plot_performance}
\alias{plot_performance}
\title{A function to perform basic portfolio analysis and plotting}
\usage{
plot_performance(x)
}
\arguments{
\item{x}{summarized output of portfolio tracker}
}
\description{
A function to perform basic portfolio analysis and plotting
}
