% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_tracker.R,
%   R/position_tracker_BACKUP_1301.R, R/position_tracker_BASE_1301.R,
%   R/position_tracker_LOCAL_1301.R, R/position_tracker_REMOTE_1301.R
\name{position_tracker}
\alias{position_tracker}
\title{A function for tracking stock positions}
\usage{
position_tracker(
  deposit = 0,
  ticker = "ticker",
  shares = 0,
  buying_price = 0,
  start_date = "yyyy-mm-dd",
  selling_price = 0,
  selling_date = "3000-01-01",
  industry = "",
  listing = c("Index", "Non-Index")
)

position_tracker(
  deposit = 0,
  ticker = "ticker",
  shares = 0,
  buying_price = 0,
  start_date = "yyyy-mm-dd",
  selling_price = 0,
  selling_date = "3000-01-01",
  industry = "",
  listing = c("Index", "Non-Index")
)

position_tracker(
  deposit = 0,
  ticker = "ticker",
  shares = 0,
  buying_price = 0,
  start_date = "yyyy-mm-dd",
  selling_price = 0,
  selling_date = "3000-01-01",
  industry = "",
  listing = c("Index", "Non-Index")
)

position_tracker(
  deposit = 0,
  ticker = "ticker",
  shares = 0,
  buying_price = 0,
  start_date = "yyyy-mm-dd",
  selling_price = 0,
  selling_date = "3000-01-01",
  industry = "",
  listing = c("Index", "Non-Index")
)

position_tracker(
  deposit = 0,
  ticker = "ticker",
  shares = 0,
  buying_price = 0,
  start_date = "yyyy-mm-dd",
  selling_price = 0,
  selling_date = "3000-01-01",
  industry = "",
  listing = c("Index", "Non-Index")
)
}
\arguments{
\item{deposit}{cash deposit}

\item{ticker}{stock ticker}

\item{shares}{number of shares bought / sold}

\item{buying_price}{buying price}

\item{start_date}{purchase date}

\item{selling_price}{selling price}

\item{selling_date}{selling date}

\item{industry}{industry type}

\item{listing}{listing type}

\item{position_tracker}{function to generate a data frame of stock position/s}
}
\description{
A function for tracking stock positions

A function for tracking stock positions

A function for tracking stock positions

A function for tracking stock positions

A function for tracking stock positions
}
