% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_historical_price.R
\name{pull_historical_price}
\alias{pull_historical_price}
\alias{historical_price_puller}
\title{Pull stock prices via pselookup API}
\usage{
pull_historical_price(
  ticker = "x",
  type = c("open", "high", "low", "close"),
  start_date = "yyyy-mm-dd",
  end_date = "yyyy-mm-dd"
)
}
\arguments{
\item{ticker}{stock ticker to pull}

\item{type}{type of price to pull}

\item{start_date}{start date of data pull}

\item{end_date}{end date of data pull}

\item{pull_historical_price}{function to pull stock prices}
}
\value{
A data frame of stock prices
}
\description{
A function to pull single stock prices listed in the Philippine Stock Exchange via pselookup API
}
