% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_multiple_prices.R
\name{pull_multiple_prices}
\alias{pull_multiple_prices}
\title{A function to generate a data frame of multiple stock prices}
\usage{
pull_multiple_prices(
  tickers,
  type = c("open", "high", "low", "close"),
  start_date = "yyyy-mm-dd",
  end_date = "yyyy-mm-dd"
)
}
\arguments{
\item{tickers}{stock tickers to pull}

\item{type}{type of price to pull}

\item{start_date}{starting date of data pull}

\item{end_date}{ending date of data pull}

\item{pull_multiple_prices}{function to pull multiple stock prices}
}
\description{
A function to pull multiple stock prices listed in the Philippine Stock Exchange via pselookup API
}
