% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_tracker.R
\name{portfolio_tracker}
\alias{portfolio_tracker}
\title{A function creating a portfolio tracker by combining a list of stocks}
\usage{
portfolio_tracker(...)
}
\arguments{
\item{...}{stock data frame to be combined; stock position output from position_tracker}
}
\description{
A function creating a portfolio tracker by combining a list of stocks
}
